Switching over to Grid Detection for mapping veins ...
Plan is to create a grid of valid points to use for pathfinding.
Player can set Leaf Point Nodes to be vein targets, then pathfinding will reach them safely from origin.
Going to use an Area2D with a Polygon Shape of the Leaf Point Nodes.
This won't include the curves, but should provide enough shape for pathfinding.